backend/requirements.txt
Flask>=2.0
spotipy>=2.19
python-dotenv>=0.19
requests>=2.25 